backend/requirements.txt
fastapi
pydantic
python-dotenv
openai
pandas
uvicorn